remove class declaration from workspace.ts
